% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandHarmonized.R
\name{calcLandHarmonized}
\alias{calcLandHarmonized}
\title{calcLandHarmonized}
\usage{
calcLandHarmonized(input, target, harmonizationPeriod, harmonization)
}
\arguments{
\item{input}{name of an input dataset, options: "magpie", "witch"}

\item{target}{name of the land target source to be used}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets.
For harmonization = "absoluteChanges" both values must be the same, the
harmonization year, which must be present in the target dataset.}

\item{harmonization}{name of harmonization method, see \code{\link{toolGetHarmonizer}}}
}
\description{
This function computes a version of the chosen land input data in the resolution
of the land input data but harmonized to the land use information of the
chosen land target data set (harmonized categories as well as harmonized
transition from historic target data to simulated input data).
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
