% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetExtrapolated.R
\name{calcLandTargetExtrapolated}
\alias{calcLandTargetExtrapolated}
\title{calcLandTargetExtrapolated}
\usage{
calcLandTargetExtrapolated(input, target, harmonizationPeriod)
}
\arguments{
\item{input}{character, name of the input data set}

\item{target}{character, name of the target data set}

\item{harmonizationPeriod}{Two integer values, will extrapolate to all years
present in input data between harmonization start and end year}
}
\value{
extrapolated land target data, if calcOutput is called with
supplementary = TRUE and target is luh2mod wood harvest area is also returned
}
\description{
Aggregated low resolution target data is extrapolated to the given years
using toolExtrapolate and normalized afterwards, so that the total sum over
all land types is unchanged.
To account for the relationship between wood harvest area and primary land
(which is no longer primary once it has been harvested) wood harvest area
is calculated here even though it is a nonland variable. The share of
woody land that was harvested in the historical period is calculated and
then multiplied by land (already extrapolated). Primary land is then
converted to secondary land, so that total reduction equals harvested area.
}
\author{
Pascal Sauer
}
