% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetLowRes.R
\name{calcLandTargetLowRes}
\alias{calcLandTargetLowRes}
\title{calcLandTargetLowRes}
\usage{
calcLandTargetLowRes(input, target, endOfHistory)
}
\arguments{
\item{input}{name of an input dataset, options: "magpie", "witch"}

\item{target}{name of a target dataset, see \code{\link{calcLandTarget}}
for available target datasets}

\item{endOfHistory}{years later than this are not returned}
}
\value{
low resolution target land data
}
\description{
Aggregate target land data to the spatial resolution of the input data in
preparation for harmonization.
}
\author{
Pascal Sauer
}
