% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHarmonized.R
\name{calcNonlandHarmonized}
\alias{calcNonlandHarmonized}
\title{calcNonlandHarmonized}
\usage{
calcNonlandHarmonized(input, target, harmonizationPeriod, harmonization)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of the target dataset}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{harmonization}{harmonization method, see \code{\link{toolGetHarmonizer}} for available methods}
}
\value{
harmonized nonland data
}
\description{
Harmonize nonland input data to target data using the specified method, checking
data for consistency before returning.
}
\details{
Wood harvest biomass (bioh) is adapted to the harmonized wood harvest area
by calculating kg C per mega hectare for input and target data and
harmonizing it. This is then multiplied by the harmonized wood harvest area
and scaled so the total harmonized bioh is reached.
Harmonize absolute fertilizer in Tg yr-1, then convert to fertilizer rate
in kg ha-1 yr-1.
}
\author{
Pascal Sauer
}
