% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandReport.R
\name{calcNonlandReport}
\alias{calcNonlandReport}
\title{calcNonlandReport}
\usage{
calcNonlandReport(outputFormat, harmonizationPeriod, yearsSubset)
}
\arguments{
\item{outputFormat}{options: ESM, ScenarioMIP}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{vector of years to keep in the output dataset}
}
\value{
nonland data
}
\description{
Convert the downscaled nonland data to the format required by the given project.
}
\examples{
\dontrun{
  calcOutput("NonlandReport", outputFormat = "ESM",
             harmonizationPeriod = c(2015, 2050), yearsSubset = 2015:2100)
}
}
\author{
Pascal Sauer
}
