% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTarget.R
\name{calcNonlandTarget}
\alias{calcNonlandTarget}
\title{calcNonlandTarget}
\usage{
calcNonlandTarget(target, endOfHistory)
}
\arguments{
\item{target}{name of a target dataset, see \code{\link{calcLandTarget}}
for available target datasets}

\item{endOfHistory}{years later than this are not returned}
}
\value{
nonland target data
}
\description{
Prepare the high resolution nonland target dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Sauer
}
