% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTargetLowRes.R
\name{calcNonlandTargetLowRes}
\alias{calcNonlandTargetLowRes}
\title{calcNonlandTargetLowRes}
\usage{
calcNonlandTargetLowRes(input, target)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2mod"}
}
\value{
low resolution target nonland data
}
\description{
Aggregate target nonland data to the spatial resolution of the input data in
preparation for harmonization. Fertilizer is converted to Tg yr-1, then aggregated and
converted back to kg ha-1 yr-1.
}
\author{
Pascal Sauer
}
