% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransitionsNC.R
\name{calcTransitionsNC}
\alias{calcTransitionsNC}
\title{calcTransitionsNC}
\usage{
calcTransitionsNC(
  outputFormat,
  harmonizationPeriod,
  yearsSubset,
  harmonization,
  downscaling
)
}
\arguments{
\item{outputFormat}{options: ESM, ScenarioMIP}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{remove years from the returned data which are not in yearsSubset}

\item{harmonization}{name of harmonization method, see \code{\link{toolGetHarmonizer}}}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}
}
\value{
data prepared to be written as a LUH-style transitions.nc file
}
\description{
Prepared data to be written as a LUH-style transitions.nc file
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
