% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWoodHarvestAreaHarmonized.R
\name{calcWoodHarvestAreaHarmonized}
\alias{calcWoodHarvestAreaHarmonized}
\title{calcWoodHarvestAreaHarmonized}
\usage{
calcWoodHarvestAreaHarmonized(
  input,
  target,
  harmonizationPeriod,
  harmonization
)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, see \code{\link{calcLandTarget}}
for available target datasets}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{harmonization}{harmonization method, see \code{\link{toolGetHarmonizer}} for available methods}
}
\value{
harmonized wood harvest area data
}
\description{
Harmonize wood harvest area based on harmonized land data.
First, wood harvest area is harmonized just like land data.
Then, to ensure consistency, primary harvest is converted to seondary
harvest (or vice versa) to match the primary land reduction from the
harmonized land data. If secondary harvest exceeds the available
secondary land, the excess is shifted to the other secondary land
(secdf to secdn and vice versa). Remaining excess harvest area is reported.
}
\examples{
\dontrun{
  calcOutput("WoodHarvestAreaHarmonized", input = "magpie",
             target = "luh2mod", harmonizationPeriod = c(2015, 2050),
             harmonization = "fade")
}
}
\author{
Pascal Sauer
}
