% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLUH3.R
\name{convertLUH3}
\alias{convertLUH3}
\title{convertLUH3}
\usage{
convertLUH3(x, subtype)
}
\arguments{
\item{x}{SpatRaster with LUH3 cell area shares}

\item{subtype}{Only "states" is converted}
}
\description{
Convert LUH3 cell area shares to absolute areas by multiplying with cell area, convert to Mha.
}
