% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLUH3.R
\name{downloadLUH3}
\alias{downloadLUH3}
\title{downloadLUH3}
\usage{
downloadLUH3(subtype)
}
\arguments{
\item{subtype}{one of states, management, transitions, cellArea}
}
\value{
metadata list with URL, title, description, author, unit, version, release date
}
\description{
Download the LUH3 dataset (states, management, transitions, static).
}
\author{
Pascal Sauer
}
