% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullSCENARIOMIP.R
\name{fullSCENARIOMIP}
\alias{fullSCENARIOMIP}
\title{fullSCENARIOMIP}
\usage{
fullSCENARIOMIP(
  rev = numeric_version("0"),
  input = "magpie",
  scenario = "",
  harmonizationPeriod = c(2025, 2050),
  yearsSubset = 1995:2100,
  harmonization = "fadeForest",
  downscaling = "magpieClassic",
  compression = 2,
  progress = TRUE
)
}
\arguments{
\item{rev}{revision number of the data. If not provided the current date will be used instead.
When called via madrat::retrieveData rev will be converted to numeric_version.}

\item{input}{name of an input dataset, options: "magpie", "witch"}

\item{scenario}{scenario name to be included in filenames}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{remove years from the returned data which are not in yearsSubset}

\item{harmonization}{name of harmonization method, see \code{\link{toolGetHarmonizer}}}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}

\item{progress}{boolean defining whether progress should be printed}
}
\description{
Run the pipeline to generate harmonized and downscaled data to report for ScenarioMIP.
LUH3 is used as historical reference dataset for harmonization and downscaling.
Write .nc files, print full report on consistency checks and write it to report.log.
}
\author{
Pascal Sauer
}
