% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMagpieFulldataGdx.R
\name{readMagpieFulldataGdx}
\alias{readMagpieFulldataGdx}
\title{readMagpieFulldataGdx}
\usage{
readMagpieFulldataGdx(subtype)
}
\arguments{
\item{subtype}{type of data to be read in. Available options are
land, crop, woodHarvestWeight, woodHarvestArea, fertilizer, clustermap}
}
\description{
Read function for data coming from the MAgPIE model.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
