% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddMetadataESM.R
\name{toolAddMetadataESM}
\alias{toolAddMetadataESM}
\title{toolAddMetadataESM}
\usage{
toolAddMetadataESM(
  ncFile,
  activityId,
  revision,
  harmonizationPeriod,
  missingValue,
  compression,
  resolution,
  references,
  targetMIP,
  ncTitle,
  referenceDataset,
  furtherInfoUrl
)
}
\arguments{
\item{ncFile}{file name of the respective nc file}

\item{activityId}{the string to store in the nc attribute activityId}

\item{revision}{the string to store in the nc attribute revision}

\item{harmonizationPeriod}{the string to store in the nc attribute harmonizationPeriod}

\item{missingValue}{the string to store in the nc attribute missingValue}

\item{compression}{the string to store in the nc attribute compression}

\item{resolution}{the string to store in the nc attribute resolution}

\item{references}{the string to store in the nc attribute references}

\item{targetMIP}{the string to store in the nc attribute targetMIP}

\item{ncTitle}{the string to store in the nc attribute ncTitle}

\item{referenceDataset}{the string to store in the nc attribute referenceDataset}

\item{furtherInfoUrl}{the string to store in the nc attribute furtherInfoUrl}
}
\description{
Add metdata to ESM compatible nc output files
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
