% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddMetadataNC.R
\name{toolAddMetadataNC}
\alias{toolAddMetadataNC}
\title{toolAddMetadataNC}
\usage{
toolAddMetadataNC(
  ncFile,
  activityId,
  revision,
  harmonizationPeriod,
  missingValue,
  compression,
  resolution,
  references,
  targetMIP,
  ncTitle,
  referenceDataset,
  furtherInfoUrl
)
}
\arguments{
\item{ncFile}{file name of the respective nc file}

\item{activityId}{a string to store in the nc attribute activityId}

\item{revision}{a string to store in the nc attribute revision}

\item{harmonizationPeriod}{a string to store in the nc attribute harmonizationPeriod}

\item{missingValue}{a string to store in the nc attribute missingValue}

\item{compression}{a string to store in the nc attribute compression}

\item{resolution}{a string to store in the nc attribute resolution}

\item{references}{a string to store in the nc attribute references}

\item{targetMIP}{a string to store in the nc attribute targetMIP}

\item{ncTitle}{a string to store in the nc attribute ncTitle}

\item{referenceDataset}{a string to store in the nc attribute referenceDataset}

\item{furtherInfoUrl}{a string to store in the nc attribute furtherInfoUrl}
}
\description{
Add metdata nc output files
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
