% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateCropland.R
\name{toolAggregateCropland}
\alias{toolAggregateCropland}
\title{toolAggregateCropland}
\usage{
toolAggregateCropland(
  land,
  cropTypes = c("c3ann", "c4ann", "c3per", "c4per", "c3nfx"),
  ...,
  keepOthers = TRUE
)
}
\arguments{
\item{land}{magpie object with variables starting with cropTypes in dim 3}

\item{cropTypes}{character vector, the variables to aggregate to}

\item{...}{reserved for future expansion}

\item{keepOthers}{logical, if FALSE cropTypes will be the only variables in output}
}
\value{
land with crop data aggregated to cropTypes
}
\description{
Aggregate variables like c3ann_irrigated and c3ann_rainfed to just c3ann.
}
\author{
Pascal Sauer
}
