% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCheckWoodHarvestArea.R
\name{toolCheckWoodHarvestArea}
\alias{toolCheckWoodHarvestArea}
\title{toolCheckWoodHarvestArea}
\usage{
toolCheckWoodHarvestArea(harvest, land, endOfHistory)
}
\arguments{
\item{harvest}{magpie object with exactly the following categories:
paste0(c("primf", "secyf", "secmf", "pltns", "primn", "secnf"), "_wood_harvest_area")}

\item{land}{magpie object with at least the following categories:
c("primf", "secdf", "pltns", "primn", "secdn")}

\item{endOfHistory}{The last year considered part of the historical period,
will check and report consistency separately for history and after}
}
\description{
Check wood harvest area is not exceeding land area of the corresponding
type divided by timestep length. Also, check that primf and primn
are reduced by at least as much as they were harvested.
}
\author{
Pascal Sauer
}
