% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDownscaleAvlCropland.R
\name{toolDownscaleAvlCropland}
\alias{toolDownscaleAvlCropland}
\title{toolDownscaleAvlCropland}
\usage{
toolDownscaleAvlCropland(x, xTarget, xTargetLowRes, mapping)
}
\arguments{
\item{x}{magclass containing land to be downscaled}

\item{xTarget}{magclass target land use dataset for initialization year}

\item{xTargetLowRes}{magclass target land use dataset for initialization year in low resolution (like x)}

\item{mapping}{mapping between spatial resolutions of \code{x} and \code{xTarget}}
}
\value{
downscaled land use dataset
}
\description{
Downscaling method using \code{luscale::interpolateAvlCroplandWeighted} as
downscaling function.
}
\author{
Pascal Sauer
}
