% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeNone.R
\name{toolHarmonizeNone}
\alias{toolHarmonizeNone}
\title{toolHarmonizeNone}
\usage{
toolHarmonizeNone(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{target data as magpie object}

\item{harmonizationPeriod}{Two integer values, until the first given
year xTarget is used, after that use xInput; the second value is ignored}
}
\value{
magpie object
}
\description{
Do not actually harmonize, but just combine xTarget and xInput,
using xTarget until start of harmonization period, then xInput.
}
\author{
Pascal Sauer
}
