% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMaxExpansion.R
\name{toolMaxExpansion}
\alias{toolMaxExpansion}
\title{toolMaxExpansion}
\usage{
toolMaxExpansion(x, ..., removeNA = FALSE)
}
\arguments{
\item{x}{A magclass object}

\item{...}{Reserved for future use}

\item{removeNA}{A logical indicating whether NA values should be ignored}
}
\value{
A numeric value indicating the maximum expansion
}
\description{
Return the maximum expansion from one timestep to the next.
}
\author{
Pascal Sauer
}
