% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMaxHarvestPerYear.R
\name{toolMaxHarvestPerYear}
\alias{toolMaxHarvestPerYear}
\title{toolMaxHarvestPerYear}
\usage{
toolMaxHarvestPerYear(land, split = TRUE)
}
\arguments{
\item{land}{magpie object with at least the following categories:
c("primf", "secdf", "pltns_added_treecover",
  "pltns_excl_added_treecover", "primn", "secdn")}

\item{split}{if TRUE: split secdf to secyf and secdmf,
rename secdf to secnf, and append "_wood_harvest_area" to names}
}
\value{
magpie object with the maximum possible yearly wood harvest area
}
\description{
Calculate the maximum possible harvest area per year
(dividing by timestep length) based on the given land data.
}
\author{
Pascal Sauer
}
