% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandReportScenarioMIP.R
\name{calcLandReportScenarioMIP}
\alias{calcLandReportScenarioMIP}
\title{calcLandReportScenarioMIP}
\usage{
calcLandReportScenarioMIP(
  harmonizationPeriod,
  yearsSubset,
  harmonization,
  downscaling
)
}
\arguments{
\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{vector of years to keep in the output dataset}

\item{harmonization}{name of harmonization method, see \code{\link{toolGetHarmonizer}}}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}
}
\value{
land use data
}
\description{
Convert the downscaled land use data to the format required by ScenarioMIP.
}
\author{
Pascal Sauer
}
