% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTarget.R
\name{calcLandTarget}
\alias{calcLandTarget}
\title{calcLandTarget}
\usage{
calcLandTarget(target)
}
\arguments{
\item{target}{name of the target dataset
luh2mod/luh3 will split secdf into pltns and secdf}
}
\value{
land target data
}
\description{
Prepare the high resolution target land use dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Sauer
}
