% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTargetExtrapolated.R
\name{calcNonlandTargetExtrapolated}
\alias{calcNonlandTargetExtrapolated}
\title{calcNonlandTargetExtrapolated}
\usage{
calcNonlandTargetExtrapolated(input, target, harmonizationPeriod)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, see \code{\link{calcLandTarget}}
for available target datasets}

\item{harmonizationPeriod}{Two integer values, will extrapolate to all years
present in input data between harmonization start and end year}
}
\value{
extrapolated nonland target data
}
\description{
Aggregated low resolution target data is extrapolated to the given years
using toolExtrapolate. To extrapolate wood harvest weight (bioh) multiply
wood harvest area already extrapolated by calcLandTargetExtrapolated with
the historical wood harvest rate in kg C per Mha. Fertilizer is extrapolated
and returned in kg ha-1 yr-1.
}
\examples{
\dontrun{
  calcOutput("NonlandTargetExtrapolated", input = "magpie",
             target = "luh3", harmonizationPeriod = c(2025, 2050))
}
}
\author{
Pascal Sauer
}
