% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStatesNC.R
\name{calcStatesNC}
\alias{calcStatesNC}
\title{calcStatesNC}
\usage{
calcStatesNC(
  outputFormat,
  input,
  harmonizationPeriod,
  yearsSubset,
  harmonization,
  downscaling
)
}
\arguments{
\item{outputFormat}{options: ESM, ScenarioMIP}

\item{input}{name of an input dataset, options: "magpie", "witch"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{remove years from the returned data which are not in yearsSubset}

\item{harmonization}{name of harmonization method, see \code{\link{toolGetHarmonizer}}}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}
}
\value{
data prepared to be written as a LUH-style states.nc file
}
\description{
Prepare data to be written as LUH-style states.nc file. Call this via calcOutput
in a full function, and set calcOutput's file argument to a .nc file path.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
