% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLUH3.R
\name{readLUH3}
\alias{readLUH3}
\title{readLUH3}
\usage{
readLUH3(subtype, subset)
}
\arguments{
\item{subtype}{one of states, management, transitions, cellArea}

\item{subset}{a vector of years that should be read}
}
\description{
Read LUH3 data. For the states subtype, the secma and secmb categories are removed.
For the management subtype, only the categories cpbf1, cpbf2, rndwd, fulwd, fertl and irrig are read.
For the transitions subtype, only the wood harvest categories bioh and harv are read. To
match magpie semantics years are shifted by 1 when reading transitions.
The LUH3 nc files have day-based time, which is converted to years.
}
