% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetHarmonizer.R
\name{toolGetHarmonizer}
\alias{toolGetHarmonizer}
\title{toolGetHarmonizer}
\usage{
toolGetHarmonizer(harmonizerName)
}
\arguments{
\item{harmonizerName}{name of a harmonizer function, currently offset, fade, fadeForest,
absoluteChanges}
}
\value{
harmonizer function
}
\description{
Get a harmonizer function by name.
}
\seealso{
\code{\link{toolHarmonizeOffset}}, \code{\link{toolHarmonizeFade}},
\code{\link{toolHarmonizeFadeForest}}, \code{\link{toolHarmonizeAbsoluteChanges}}
}
\author{
Pascal Sauer
}
