% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeAbsoluteChanges.R
\name{toolHarmonizeAbsoluteChanges}
\alias{toolHarmonizeAbsoluteChanges}
\title{toolHarmonizeAbsoluteChanges}
\usage{
toolHarmonizeAbsoluteChanges(xInput, xTarget, harmonizationPeriod, level = 1)
}
\arguments{
\item{xInput}{magpie object with the data to take absolute changes from,
usually model projections}

\item{xTarget}{magpie object with the historical target data}

\item{harmonizationPeriod}{a single integer year that must exist in both
xInput and xTarget; before this year the target dataset is used, from this
year onward the target value at harmonizationPeriod plus the absolute
change of the input data (relative to harmonizationPeriod) is used}

\item{level}{passed to toolReplaceExpansion}
}
\value{
magpie object with harmonized data, negative values are clamped to
0 and each region/timestep is renormalized to keep the total area constant,
emitting a warning if any clamping happened
}
\description{
Harmonize datasets by applying the absolute changes of the input data
relative to a single harmonization year on top of the target data at that
year. This is the aneris/mip "offset" method without convergence, i.e. a
permanent offset that is never faded out.
}
\author{
Pascal Sauer
}
