% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeAbsoluteChanges.R
\name{toolHarmonizeAbsoluteChanges}
\alias{toolHarmonizeAbsoluteChanges}
\title{toolHarmonizeAbsoluteChanges}
\usage{
toolHarmonizeAbsoluteChanges(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{target data as magpie object}

\item{harmonizationPeriod}{Single integer value, the year the absolute
changes of the input data are applied to, must be present in both input and
target data}
}
\value{
harmonized data set as magpie object with data from target for years
up to and including the harmonization year and absolute changes from input
relative to the harmonization year afterwards
}
\description{
Tool function for creating a harmonized data set by applying the absolute
changes of the input data to the target data: up to the harmonization year
the target data is used, afterwards the difference of the input data to the
input data of the harmonization year is added to the target data of the
harmonization year.
If this makes forest categories (pltns, primf, secdf) negative, the missing
area is deducted proportionally from the other land categories (primn,
secdn). Remaining negative values are set to 0 and categories are scaled
down to keep the total area constant: first everything except
prim (primf, primn) and urban, then prim only if prim + urban exceed
total area. Urban is never scaled.
}
\author{
Pascal Sauer
}
