% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeAbsoluteChanges.R
\name{toolHarmonizeAbsoluteChanges}
\alias{toolHarmonizeAbsoluteChanges}
\title{toolHarmonizeAbsoluteChanges}
\usage{
toolHarmonizeAbsoluteChanges(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{target data as magpie object}

\item{harmonizationPeriod}{Single integer value, the year the absolute
changes of the input data are applied to, must be present in both input and
target data}
}
\value{
harmonized data set as magpie object with data from target for years
up to and including the harmonization year and absolute changes from input
relative to the harmonization year afterwards
}
\description{
Tool function for creating a harmonized data set by applying the absolute
changes of the input data to the target data. Before and in the
harmonization year the target data is used. Afterwards, for every year the
difference of the input data to the input data in the harmonization year is
added to the target data of the harmonization year, e.g. if the target data
has 10 Mha secdf in the harmonization year and the input data secdf grows by
2 Mha from the harmonization year to 2025, the harmonized data set has 12
Mha secdf in 2025.
If this makes values negative, they are set to 0 and the other categories
are reduced proportionally, so that the total area stays constant.
}
\author{
Pascal Sauer
}
