% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeFadeForest.R
\name{toolHarmonizeFadeForest}
\alias{toolHarmonizeFadeForest}
\title{toolHarmonizeFadeForest}
\usage{
toolHarmonizeFadeForest(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{target data as magpie object}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
harmonized data set as magpie object with data from input for years
before the harmonization period, data from target for years after the
harmonization period and a smooth transition in between.
}
\description{
Harmonize using \code{\link{toolHarmonizeFade}}, then sum up primf + secdf,
then disaggregate again in a way that minimizes primf to secdf conversion.
}
\author{
Pascal Sauer
}
