% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeNone.R
\name{toolHarmonizeNone}
\alias{toolHarmonizeNone}
\title{toolHarmonizeNone}
\usage{
toolHarmonizeNone(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{ignored for this function}

\item{harmonizationPeriod}{ignored for this function}
}
\value{
unchanged xInput
}
\description{
Do not actually harmonize, just return xInput.
}
\author{
Pascal Sauer
}
